requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.2 